gcc16 merge

+++ binutils/gprofng/testsuite/gprofng.display/synprog/endcases.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2021 Free Software Foundation, Inc.
+/* Copyright (C) 2021-2026 Free Software Foundation, Inc.
    Contributed by Oracle.
 
    This file is part of GNU Binutils.
@@ -40,8 +40,8 @@ static void s_inline_code (int);
 void ext_inline_code (int);
 
 #ifndef NO_INLINE
-void xinline_code () __attribute__ ((always_inline));
-void s_inline_code () __attribute__ ((always_inline));
+void xinline_code (int) __attribute__ ((always_inline));
+void s_inline_code (int) __attribute__ ((always_inline));
 #endif
 
 #include "inc_inline.h"
@@ -52,7 +52,7 @@ int x2M = 2000000;
 int x8M = 8000000;
 
 /* define a macro that burns CPU time */
-#define burncpu(nn) \ 
+#define burncpu(nn) \
         x = 0; \
         for (j = 0; j < (nn * x8M); j++) { \
                  x = x + 1; \
